Kleinere Anpassung an Code. - Kontrolle ob ein Pixel außerhalb des Bildes liegt hinzugefügt
BUG:
- Compiler sagt, dass es einen Vergleich mit einem Point gäbe und dies nicht definiert sei

+++ CV2_Excersice1/SimilarityMeasure.h
@@ -3,6 +3,6 @@
 
 #include <opencv2/core/core.hpp>
 
-double colorSSD(const cv::Mat* firstImage, const cv::Mat* secondImage, cv::Point imagePoint, int windowSize, cv::Point offset);
+long colorSSD(const cv::Mat* firstImage, const cv::Mat* secondImage, cv::Point imagePoint, int windowSize, cv::Point offset);
 
 #endif // !SIMILARITYMEASURE_H
